web-animations/animation-player/animation-player-paused.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Animation player paused attribute test</title>
<meta name="assert" content="A boolean value that is true if the player is currently paused. The paused state is initially false.">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#widl-AnimationPlayer-paused">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
function newAnimation(animationTarget) {
    animationTarget.style.width = '300px';
    return new Animation(animationTarget, [
        {width: '10px', offset: 0},
        {width: '100px', offset: 1/2},
        {width: '200px', offset: 1}], 1000);
}

function newAnimationPlayer(animationTarget) {
    var animation = newAnimation(animationTarget);
    return document.timeline.play(animation);
}


test(function() {
    var player = newAnimationPlayer(createDiv(this));

    assert_false(player.paused, 'The value of AnimationPlayer.paused attribute should be ' +
        'initially false');
}, 'The value of AnimationPlayer.paused attribute should be initially false');


test(function() {
    var player = newAnimationPlayer(createDiv(this));

    player.currentTime = 999;
    assert_false(player.paused, 'The value of AnimationPlayer.paused attribute should be ' +
        'false if player is not paused');

    player.currentTime = 1001;
    assert_false(player.paused, 'The value of AnimationPlayer.paused attribute should be ' +
        'false if player is not paused but seek time is greater than content end');
}, 'The value of AnimationPlayer.paused attribute should be false if player is not paused');


test(function() {
    var player = newAnimationPlayer(createDiv(this));

    player.pause();
    assert_true(player.paused, 'AnimationPlayer should be paused');

    player.play();
    assert_false(player.paused, 'AnimationPlayer should not be paused');

    player.pause();
    assert_true(player.paused, 'AnimationPlayer should be paused again');
}, 'Test true value of AnimationPlayer.paused attribute');


async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    player.currentTime = 990;
    setTimeout(function() {
        t.step(function () {
            assert_false(player.paused, 'The value of AnimationPlayer.paused attribute ' +
                'should be false');
            player.pause();
            assert_true(player.paused, 'The value of AnimationPlayer.paused attribute ' +
                'should be true');
            t.done();
        });
    }, 50);
}, 'Test AnimationPlayer.paused attribute when player is finished');


async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    player.currentTime = 10;
    player.playbackrate = -1;
    setTimeout(function() {
        t.step(function () {
            assert_false(player.paused, 'The value of AnimationPlayer.paused attribute ' +
                'should be false');
            player.pause();
            assert_true(player.paused, 'The value of AnimationPlayer.paused attribute ' +
                'should be true');
            t.done();
        });
    }, 30);
}, 'Test AnimationPlayer.paused attribute when player is finished and playback rate < 0');


test(function() {
    var animation = newAnimation(createDiv(this))
    var player = document.timeline.play(animation);
    player.currentTime = 900;
    player.pause();
    animation.timing.playbackRate = 2;
    assert_true(player.paused, 'The value of AnimationPlayer.paused attribute ' +
        'should be true');
}, 'Test AnimationPlayer.paused attribute when animation length is changed (shortened)');


test(function() {
    var animation = newAnimation(createDiv(this))
    var player = document.timeline.play(animation);
    player.currentTime = 1100;
    player.pause();
    animation.timing.playbackRate = 0.5;
    assert_true(player.paused, 'The value of AnimationPlayer.paused attribute ' +
        'should be true');
}, 'Test AnimationPlayer.paused attribute when animation length is changed (extended)');
</script>
</body>
